+++ python/sina/__init__.py
@@ -1,7 +1,7 @@
 """Logic for handling inserting, extracting, and transforming data between backends."""
 
 # Note that the C++ spack package builder references this version string
-__VERSION__ = "1.10.0"
+__VERSION__ = "1.11.0"
 
 
 from sina.datastore import connect