+++ update-version.sh
@@ -3,7 +3,7 @@
 set -e
 
 MAJOR=1
-MINOR=10
+MINOR=11
 PATCH=0
 FULL=${MAJOR}.${MINOR}.${PATCH}